refactor(app): colocate diff-render cluster under git/diff-file-body/
