+++ packages/app/src/git/diff-file-body/highlighted-text.tsx
